New class for Emme-scenario
* New classes that hold Emme-scenario specific methods. This makes it easier to assign single time periods and to do extended analyses.
* No modifications needed to modelsystem. Only Emme-assignment is changed.
